% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness-measures.R
\name{.second_moa_ring}
\alias{.second_moa_ring}
\title{Compute the second moment of area for a single ring}
\usage{
.second_moa_ring(coords, centroid)
}
\description{
Compute the second moment of area for a single ring
}
