% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness-measures.R
\name{.second_moment_single_geom}
\alias{.second_moment_single_geom}
\title{Second areal moment for a single geometry}
\usage{
.second_moment_single_geom(geom)
}
\arguments{
\item{geom}{A simple feature geometry}
}
\value{
A single number, giving the second areal moment for the geometry
}
\description{
Computes the second areal moment for a single geometry (POLYGON or MULTIPOLYGON)
}
\keyword{internal}
