% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_connected_components}
\alias{rust_connected_components}
\title{Find connected components}
\usage{
rust_connected_components(i, j, n)
}
\arguments{
\item{i}{Row indices of adjacency matrix}

\item{j}{Column indices of adjacency matrix}

\item{n}{Number of nodes}
}
\value{
Vector of component labels (0-based)
}
\description{
Find connected components
}
