% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_corridor}
\alias{rust_corridor}
\title{Least-cost corridor routing on a raster grid}
\usage{
rust_corridor(
  values,
  n_rows,
  n_cols,
  cell_width,
  cell_height,
  origin_cell,
  dest_cell,
  neighbours,
  method
)
}
\arguments{
\item{values}{Flattened cell values (row-major), NaN = impassable}

\item{n_rows}{Number of rows in the raster}

\item{n_cols}{Number of columns in the raster}

\item{cell_width}{Cell width in CRS units}

\item{cell_height}{Cell height in CRS units}

\item{origin_cell}{0-based cell index of origin}

\item{dest_cell}{0-based cell index of destination}

\item{neighbours}{Cell connectivity: 4, 8, or 16}

\item{method}{Routing algorithm: "dijkstra", "bidirectional", or "astar"}
}
\value{
List with path_cells, total_cost, solve_time_ms, graph_build_time_ms, n_edges
}
\description{
Least-cost corridor routing on a raster grid
}
