% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_corridor_build_graph}
\alias{rust_corridor_build_graph}
\title{Build a cached corridor graph from a raster grid}
\usage{
rust_corridor_build_graph(
  values,
  n_rows,
  n_cols,
  cell_width,
  cell_height,
  neighbours
)
}
\description{
Build a cached corridor graph from a raster grid
}
