% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_corridor_graph_info}
\alias{rust_corridor_graph_info}
\title{Get metadata from a cached corridor graph}
\usage{
rust_corridor_graph_info(graph)
}
\description{
Get metadata from a cached corridor graph
}
