% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_corridor_solve_cached}
\alias{rust_corridor_solve_cached}
\title{Solve on a cached corridor graph}
\usage{
rust_corridor_solve_cached(graph, origin_cell, dest_cell, method)
}
\description{
Solve on a cached corridor graph
}
