% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_distance_matrix_euclidean}
\alias{rust_distance_matrix_euclidean}
\title{Compute Euclidean distance matrix (Rust)}
\usage{
rust_distance_matrix_euclidean(x1, y1, x2, y2)
}
\description{
Compute Euclidean distance matrix (Rust)
}
\keyword{internal}
