% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_is_connected}
\alias{rust_is_connected}
\title{Check if a graph is connected}
\usage{
rust_is_connected(i, j, n)
}
\arguments{
\item{i}{Row indices of adjacency matrix}

\item{j}{Column indices of adjacency matrix}

\item{n}{Number of nodes}
}
\value{
TRUE if connected, FALSE otherwise
}
\description{
Check if a graph is connected
}
