% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_lscp}
\alias{rust_lscp}
\title{Solve LSCP (Location Set Covering Problem)}
\usage{
rust_lscp(cost_matrix, service_radius)
}
\arguments{
\item{cost_matrix}{Cost/distance matrix (demand x facilities)}

\item{service_radius}{Maximum service distance}
}
\value{
List with selected facilities and coverage
}
\description{
Solve LSCP (Location Set Covering Problem)
}
