% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_p_dispersion}
\alias{rust_p_dispersion}
\title{Solve P-Dispersion problem (Rust)}
\usage{
rust_p_dispersion(distance_matrix, n_facilities)
}
\description{
Solve P-Dispersion problem (Rust)
}
\keyword{internal}
