% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_p_median}
\alias{rust_p_median}
\title{Solve P-Median facility location problem}
\usage{
rust_p_median(
  cost_matrix,
  weights,
  n_facilities,
  fixed_facilities,
  max_distance
)
}
\arguments{
\item{cost_matrix}{Cost/distance matrix (demand x facilities)}

\item{weights}{Demand weights}

\item{n_facilities}{Number of facilities to locate (p)}

\item{fixed_facilities}{Optional indices of pre-selected facilities (1-based)}

\item{max_distance}{Optional maximum distance for assignments}
}
\value{
List with selected facilities and assignments
}
\description{
Solve P-Median facility location problem
}
