% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_skater}
\alias{rust_skater}
\title{Solve SKATER regionalization (Rust)}
\usage{
rust_skater(attrs, adj_i, adj_j, n_regions, floor_var, floor_value, seed)
}
\description{
Solve SKATER regionalization (Rust)
}
\keyword{internal}
