% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_spenc}
\alias{rust_spenc}
\title{Solve SPENC regionalization problem (Rust)}
\usage{
rust_spenc(attrs, n_regions, adj_i, adj_j, gamma, seed)
}
\description{
Solve SPENC regionalization problem (Rust)
}
\keyword{internal}
