% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_ward_constrained}
\alias{rust_ward_constrained}
\title{Solve spatially-constrained Ward clustering (Rust)}
\usage{
rust_ward_constrained(attrs, n_regions, adj_i, adj_j)
}
\description{
Solve spatially-constrained Ward clustering (Rust)
}
\keyword{internal}
