% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness-measures.R
\name{.second_moment_single_geom}
\alias{.second_moment_single_geom}
\title{Process a single geometry (handles POLYGON and MULTIPOLYGON)}
\usage{
.second_moment_single_geom(geom)
}
\description{
Process a single geometry (handles POLYGON and MULTIPOLYGON)
}
