% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_azp}
\alias{rust_azp}
\title{Solve AZP regionalization problem (Rust)}
\usage{
rust_azp(
  attrs,
  n_regions,
  adj_i,
  adj_j,
  method,
  max_iterations,
  tabu_length,
  cooling_rate,
  initial_temperature,
  seed
)
}
\description{
Solve AZP regionalization problem (Rust)
}
\keyword{internal}
