% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_cflp}
\alias{rust_cflp}
\title{Solve Capacitated Facility Location Problem (CFLP)}
\usage{
rust_cflp(cost_matrix, weights, capacities, n_facilities, facility_costs)
}
\arguments{
\item{cost_matrix}{Cost/distance matrix (demand x facilities)}

\item{weights}{Demand weights}

\item{capacities}{Capacity of each facility}

\item{n_facilities}{Number of facilities to locate (0 if using facility costs)}

\item{facility_costs}{Optional fixed cost to open each facility}
}
\value{
List with selected facilities, assignments, utilizations
}
\description{
Minimize weighted distance subject to facility capacity constraints.
Supports fixed number of facilities or facility opening costs.
}
