% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_cflp}
\alias{rust_cflp}
\title{Solve CFLP (Capacitated Facility Location Problem) (Rust)}
\usage{
rust_cflp(cost_matrix, weights, capacities, n_facilities, facility_costs)
}
\description{
Solve CFLP (Capacitated Facility Location Problem) (Rust)
}
\keyword{internal}
