% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_connected_components}
\alias{rust_connected_components}
\title{Find connected components (Rust)}
\usage{
rust_connected_components(i, j, n)
}
\description{
Find connected components (Rust)
}
\keyword{internal}
