% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_distance_matrix_manhattan}
\alias{rust_distance_matrix_manhattan}
\title{Compute Manhattan distance matrix (Rust)}
\usage{
rust_distance_matrix_manhattan(x1, y1, x2, y2)
}
\description{
Compute Manhattan distance matrix (Rust)
}
\keyword{internal}
