% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_frlm_greedy}
\alias{rust_frlm_greedy}
\title{Solve FRLM using greedy heuristic (Rust)}
\usage{
rust_frlm_greedy(
  n_candidates,
  path_candidates,
  path_offsets,
  path_distances,
  flow_volumes,
  vehicle_range,
  n_facilities
)
}
\description{
Solve FRLM using greedy heuristic (Rust)
}
\keyword{internal}
