% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_huff}
\alias{rust_huff}
\title{Compute Huff Model probabilities (Rust)}
\usage{
rust_huff(cost_matrix, attractiveness, distance_exponent, sales_potential)
}
\description{
Compute Huff Model probabilities (Rust)
}
\keyword{internal}
