% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_max_p}
\alias{rust_max_p}
\title{Solve Max-P regionalization problem (Rust)}
\usage{
rust_max_p(
  attrs,
  threshold_var,
  threshold,
  adj_i,
  adj_j,
  n_iterations,
  n_sa_iterations,
  cooling_rate,
  tabu_length,
  seed,
  centroids_x,
  centroids_y,
  compact,
  compact_weight
)
}
\description{
Solve Max-P regionalization problem (Rust)
}
\keyword{internal}
