% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_p_median}
\alias{rust_p_median}
\title{Solve P-Median problem (Rust)}
\usage{
rust_p_median(cost_matrix, weights, n_facilities)
}
\description{
Solve P-Median problem (Rust)
}
\keyword{internal}
