% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spopt-package.R
\docType{package}
\name{spopt-package}
\alias{spopt}
\alias{spopt-package}
\title{spopt: Spatial Optimization for Regionalization, Facility Location, and Market Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements spatial optimization algorithms across several problem families: contiguity-constrained regionalization, discrete facility location, market share analysis, and least-cost corridor and route optimization over raster cost surfaces. Facility location problems also accept user-supplied network travel-time matrices. Uses a 'Rust' backend via 'extendr' for graph and routing algorithms, and the 'HiGHS' solver via the 'highs' package for facility location mixed-integer programs. Method-level references are provided in the documentation of the individual functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://walker-data.com/spopt-r/}
  \item \url{https://github.com/walkerke/spopt-r}
  \item Report bugs at \url{https://github.com/walkerke/spopt-r/issues}
}

}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle@walker-data.com}

Other contributors:
\itemize{
  \item PySAL Developers (Original Python spopt library) [copyright holder]
}

}
\keyword{internal}
